ad resources classes

+++ database/migrations/2022_07_11_134015_create_ads_table.php
@@ -17,7 +17,7 @@ class CreateAdsTable extends Migration
             $table->bigIncrements('id');
             $table->string('name','200');
             $table->string('description','1000')->nullable();
-            $table->float('price');
+            $table->unsignedInteger('price');
             $table->timestamps();
         });
     }
